requirements.txt
absl-py==0.12.0
appnope==0.1.2
asgiref==3.3.4
astunparse==1.6.3
attrs==20.3.0
audioread==2.1.9
backcall==0.2.0
bokeh==2.3.1
bz2file==0.98
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
decorator==4.4.2
dill==0.3.3
Django==3.2
djangorestframework==3.12.4
dm-sonnet==2.0.0
dm-tree==0.1.6
dopamine-rl==3.0.1
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.1.2
gin-config==0.4.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.32.0
gunicorn==20.1.0
gym==0.18.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.1
importlib-resources==5.1.2
intervaltree==3.1.0
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
Keras-Preprocessing==1.1.2
kfac==0.2.0
kiwisolver==1.3.1
librosa==0.7.2
llvmlite==0.32.1
magenta==2.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mesh-tensorflow==0.1.19
mido==1.2.6
mir-eval==0.6
mpmath==1.2.1
networkx==2.5.1
note-seq==0.0.3
numba==0.49.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pretty-midi==0.2.9
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
pyglet==1.5.0
Pygments==2.8.1
pygtrie==2.4.2
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
python-rtmidi==1.1.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sk-video==1.1.10
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
sox==1.4.1
sqlparse==0.4.1
sympy==1.8
tabulate==0.8.9
tensor2tensor==1.15.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.4.0
tensorflow-gan==2.0.0
tensorflow-hub==0.12.0
tensorflow-metadata==0.29.0
tensorflow-probability==0.7.0
termcolor==1.1.0
tf-slim==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typeguard==2.12.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
